Moved changes made in the new student service to the existing student service and cleaned up the loadAndUpdateStudent method in the new database updater.
