+++ src/services/doo_api_database_updater/database_updater.js
@@ -1,5 +1,5 @@
 const Oodi = require('./oodi_interface')
-const StudentService = require('./students')
+const StudentService = require('../students')
 const logger = require('../../util/logger')
 
 process.on('unhandledRejection', (reason) => {
@@ -7,55 +7,36 @@ process.on('unhandledRejection', (reason) => {
 })
 
 const updateStudentInformation = async studentList => {
-  const students = await (studentList.map(student => loadAndUpdateStudent(student)))
-}
+  await Promise.all(studentList.map(student => loadAndUpdateStudent(student)))
+} 
+
 const loadAndUpdateStudent = async studentNumber => {
   try {
-    let [ student, studentFromOodi ] = await Promise.all([StudentService.byId(studentNumber), Oodi.getStudent(studentNumber)
+    let [ studentFromDb, studentFromApi ] = await Promise.all([StudentService.byId(studentNumber), Oodi.getStudent(studentNumber)
     ])
     
-    if (studentFromOodi === null) {
+    if (studentFromApi === null) {
       return null
     }
 
-    if (student === null) {
+    if (studentFromDb === null) {
       try {
-        student = await StudentService.createStudent(studentFromOodi)
+        studentFromDb = await StudentService.createStudent(studentFromApi)
         logger.verbose('Student ' + studentNumber + ' created to database')
-        console.log(student)
-        return [student, studentFromOodi]
+        return [studentFromDb, studentFromApi]
       } catch (e) {
         logger.error('Student ' + studentNumber + ': creation failed, error message:')
         return null
       }
     }
-
-  //   const oodiLastCreditDate = studentFromOodi[21] !== null ?
-  //     getDate(studentFromOodi[21], 'DD.MM.YYYY') : null
-
-  //   if (oodiLastCreditDate === null ||
-  //     oodiLastCreditDate === getDate(student.dataValues.dateoflastcredit, 'YYYY-MM-DD')) {
-  //     return student
-  //   }
-
-  //   try {
-  //     await StudentService.updateStudent(studentFromOodi)
-  //     logger.verbose('Student ' + studentNumber + ': details updated')
-  //   } catch (e) {
-  //     logger.error('Student ' + studentNumber + ': update failed')
-  //   }
-
-  //   return student
   } catch (e) {
-  //   logger.error('Student: ' + studentNumber + ' loadAndUpdate failed')
-  // }
+    logger.error('Student: ' + studentNumber + ' loadAndUpdate failed')
   }
 }
 
 const run = async () => {
   const studentList = ['014441008', '014420676', '014580736']
-  const [studyrights, credits] = await updateStudentInformation(studentList)
-
+  await updateStudentInformation(studentList)
   process.exit(0)
 }
